Removed global CURLSetup, made it part of Context
