+++ CMakeLists.txt
@@ -109,6 +109,8 @@ set(LIBPOWERLOADER_SRCS
     ${POWERLOADER_SOURCE_DIR}/target.cpp
     ${POWERLOADER_SOURCE_DIR}/url.cpp
     ${POWERLOADER_SOURCE_DIR}/utils.cpp
+    ${POWERLOADER_SOURCE_DIR}/target.hpp
+    ${POWERLOADER_SOURCE_DIR}/curl_internal.hpp
 )
 if (WITH_ZCHUNK)
     list(APPEND LIBPOWERLOADER_SRCS ${POWERLOADER_SOURCE_DIR}/zck.cpp
